+++ WebApps/database/migrations/2017_07_22_155657_create_product_table.php
@@ -13,28 +13,31 @@ class CreateProductTable extends Migration
      */
     public function up()
     {
+        if (!Schema::hasTable('products')) {
 
-        Schema::dropIfExists('products');
+            Schema::dropIfExists('products');
 
-        Schema::create('products', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('name');
-            $table->double('weight');
-            $table->string('weight_type');
-            $table->boolean('active');
-            $table->text('relatedNames');
-            $table->timestamps();
+            Schema::create('products', function (Blueprint $table) {
+                $table->increments('id');
+                $table->string('name');
+                $table->double('weight');
+                $table->string('weight_type');
+                $table->boolean('active');
+                $table->text('relatedNames');
+                $table->timestamps();
 
-            $table->integer('brand_id')->unsigned()->nullable();
-            $table->foreign('brand_id')
-            ->references('id')->on('brands')->onDelete('cascade');
+                $table->integer('brand_id')->unsigned()->nullable();
+                $table->foreign('brand_id')
+                ->references('id')->on('brands')->onDelete('cascade');
 
-            $table->integer('sub_category_id')->unsigned()->nullable();
-            $table->foreign('sub_category_id')
-            ->references('id')->on('sub_categories')->onDelete('cascade');
-        });
+                $table->integer('sub_category_id')->unsigned()->nullable();
+                $table->foreign('sub_category_id')
+                ->references('id')->on('sub_categories')->onDelete('cascade');
+            });
 
-        DB::statement('ALTER TABLE products ADD FULLTEXT search(name)');
+            DB::statement('ALTER TABLE products ADD FULLTEXT search(name)');
+            
+        }
     }
 
     /**
