+++ WebApps/database/migrations/2017_07_23_192608_create_product_retailer_table.php
@@ -13,27 +13,29 @@ class CreateProductRetailerTable extends Migration
      */
     public function up()
     {
-        Schema::create('product_retailers', function (Blueprint $table) {
-            $table->increments('id');
-            $table->double('price');
-            $table->double('base_price');
-            $table->boolean('hasDiscount');
-            $table->double('price_per_weight');
-            $table->string('type_of_weight');
-            $table->integer('pid');
-            $table->text('image');
-            $table->text('link');
-            $table->boolean('active');
-            $table->timestamps();
+        if (!Schema::hasTable('product_retailers')) {
+            Schema::create('product_retailers', function (Blueprint $table) {
+                $table->increments('id');
+                $table->double('price');
+                $table->double('base_price');
+                $table->boolean('hasDiscount');
+                $table->double('price_per_weight');
+                $table->string('type_of_weight');
+                $table->integer('pid');
+                $table->text('image');
+                $table->text('link');
+                $table->boolean('active');
+                $table->timestamps();
 
-            $table->integer('product_id')->unsigned()->nullable();
-            $table->foreign('product_id')
-            ->references('id')->on('products')->onDelete('cascade');
+                $table->integer('product_id')->unsigned()->nullable();
+                $table->foreign('product_id')
+                ->references('id')->on('products')->onDelete('cascade');
 
-            $table->integer('retailer_id')->unsigned()->nullable();
-            $table->foreign('retailer_id')
-            ->references('id')->on('retailers')->onDelete('cascade');
-        });
+                $table->integer('retailer_id')->unsigned()->nullable();
+                $table->foreign('retailer_id')
+                ->references('id')->on('retailers')->onDelete('cascade');
+            });
+        }
     }
 
     /**
